requirements.txt
numpy
click==6.7
attrs==15.2.0
pandas
opencv-python
scikit-learn
matplotlib
moviepy
haikunator